Added dryrun option

+++ curry_batch/__main__.py
@@ -10,19 +10,23 @@ def worker_init():
     signal.signal(signal.SIGINT, signal.SIG_IGN)
 
 def execute_command(args):
-    command, arg_list = args
+    command, arg_list, dryrun = args
     wc_dict = enumerate_wildcards(arg_list)
     try:
         formatted_command = safe_format(command, wc_dict)
     except Exception as e:
         print(f"Error in execute_command for arg_list {arg_list} and command template {command}: {e}")
-        return None
-    result = subprocess.run(formatted_command, shell=True, capture_output=True, text=True)
-    return result.stdout.strip()
+    if dryrun:
+        click.echo(formatted_command)
+        result = ''
+    else:
+        result = subprocess.run(formatted_command, shell=True, capture_output=True, text=True)
+        result = result.stdout.strip()
+    return result
 
-def execute_commands_multiprocessing(commands, arguments, max_processes=None):
+def execute_commands_multiprocessing(commands, arguments, max_processes=None, dryrun=False):
     output = []
-    args_for_pool = [(command, args) for args in arguments for command in commands]
+    args_for_pool = [(command, args, dryrun) for args in arguments for command in commands]
 
     with Pool(processes=max_processes, initializer=worker_init) as pool:
         try:
@@ -42,7 +46,7 @@ def execute_commands_multiprocessing(commands, arguments, max_processes=None):
 
 def safe_format(template, kwargs):
     for wc, arg in kwargs.items():
-        template = template.replace(wc, arg)
+        template = template.replace(wc, str(arg))
     return template
 
 def enumerate_wildcards(arguments):
@@ -54,12 +58,14 @@ def enumerate_wildcards(arguments):
 
 @click.command()
 @click.argument('commands', nargs=-1, required=True)
-@click.option('--max-processes', default=1, help='Maximum number of processes to use for parallel execution')
-def curry_batch(commands, max_processes):
+@click.option('--max-processes', default=None, help='Limit number of processes to use for parallel execution')
+@click.option('--dryrun', is_flag = True, help='Echo commands to stdout rather than running them')
+def curry_batch(commands, max_processes, dryrun):
     input = sys.stdin.read()
     arguments = yaml.safe_load(input)
-    output = execute_commands_multiprocessing(commands, arguments, max_processes)
-    click.echo(yaml.dump(output))
+    output = execute_commands_multiprocessing(commands, arguments, max_processes, dryrun)
+    if not dryrun:
+        click.echo(yaml.dump(output))
 
 if __name__ == '__main__':
     curry_batch()
